requirements.txt
ollama
django
djangorestframework
